Deepworld
This is the proof of concept of the mod, not much can be done but explore the unfinished mines. Only 20% of content is in game. Many recipes are missing, some loot docent spawn, enemy density is to high (Testing porpuse), Many mines can be found at a time because of very high weight value (For testing the gen).

Alot is not done but wanted to upload to prevent data lost and as well any ideas or errors found